refactor(server): /api/doctor 라우트 분할 (Plan C #C9 8단계)
- src/server/routes/doctor.ts 신규: 4개 라우트 캡슐화
  (run, heal/stdin, heal/:id Level1, heal/:id/stream Level2)
- safeError 적용 (1회) + Level1 헬퍼는 sanitize 단순화
- dashboard-api.ts에서 doctor 모듈 import 정리 (runAllChecks/healLevel1/healLevel2/
  writeToActiveHealProcess)
- dashboard-api.ts 1058 → 983줄 (-75) — **1000줄 아래로 진입**

남은 도메인: repositories/health/sse/skip-events/auth/new-issue/projects-health
